Bind textures to MiniMapWindow
NOTE: This currently doesn't work properly. I didn't realize I can only draw textures using the Renderer it was generated with. I'll need to generate the textures inside the MiniMapWindow itself or get the Renderer that the MMW uses out to the textures (ew). Not enough time to do it tonight so leaving this note here.